omg this literally drove me crazy
idk what the problem is, but the sparse least squares method must just
ignore a ton of the entries. idk maybe there is a way to get it to work,
but I couldn't figure it out. the dense version solved pretty much all
my problems and I lived happily again